fix: refactor listFiles and listFilesByPrefix samples into their own files
